alteracoes

+++ src/com/company/Route.java
@@ -4,13 +4,13 @@ public class Route {
     private StartPoint startPoint;
     private EndPoint endPoint;
 
-    public Route(StartPoint startPoint, EndPoint endPoint) {
+    public Route(StartPoint startPoint, EndPoint endPoint){
 
         this.startPoint=startPoint;
         this.endPoint=endPoint;
     }
 
-    public double getDistance() {
+    public int getDistance() {
         int distx = this.endPoint.getCoordinateX() - this.startPoint.getCoordinateX();
         distx=Math.abs(distx);
         int disty=this.endPoint.getCoordinateY()-this.startPoint.getCoordinateY();
